feat(benchmarks): share streaming TTFT/TTFA metrics across MMMU and MMSU
