+++ benchmarks/tasks/visual_understand.py
@@ -16,7 +16,7 @@ import aiohttp
 
 from benchmarks.benchmarker.data import RequestResult
 from benchmarks.benchmarker.runner import SendFn
-from benchmarks.benchmarker.utils import parse_sse_event
+from benchmarks.benchmarker.utils import read_streaming_chat_response
 from benchmarks.dataset.mmmu import MMMUSample
 
 logger = logging.getLogger(__name__)
@@ -270,7 +270,7 @@ def make_mmmu_send_fn(
             async with session.post(api_url, json=payload) as response:
                 response.raise_for_status()
                 if stream:
-                    usage = await _read_streaming_mmmu_response(
+                    usage = await read_streaming_chat_response(
                         response,
                         result,
                         start_time=start_time,
@@ -321,73 +321,6 @@ def make_mmmu_send_fn(
         return result
 
     return send_fn
-
-
-async def _read_streaming_mmmu_response(
-    response: aiohttp.ClientResponse,
-    result: RequestResult,
-    *,
-    start_time: float,
-    expect_audio: bool,
-) -> dict:
-    """Consume SSE deltas for streaming MMMU performance measurement."""
-    text_parts: list[str] = []
-    usage: dict = {}
-    last_audio_time: float | None = None
-    buffer = bytearray()
-
-    def consume_line(raw_line: bytes) -> None:
-        nonlocal last_audio_time, usage
-        try:
-            event = parse_sse_event(raw_line.decode("utf-8", errors="replace").strip())
-        except ValueError:
-            return
-        if event is None:
-            return
-
-        event_usage = event.get("usage")
-        if isinstance(event_usage, dict):
-            usage = event_usage
-
-        for choice in event.get("choices", []):
-            if not isinstance(choice, dict):
-                continue
-            delta = choice.get("delta")
-            if not isinstance(delta, dict):
-                continue
-
-            content = delta.get("content")
-            if isinstance(content, str) and content:
-                if result.text_ttft_s is None:
-                    result.text_ttft_s = time.perf_counter() - start_time
-                text_parts.append(content)
-
-            audio = delta.get("audio")
-            if not isinstance(audio, dict) or not isinstance(audio.get("data"), str):
-                continue
-            now = time.perf_counter()
-            if result.audio_ttfp_s is None:
-                result.audio_ttfp_s = now - start_time
-            elif last_audio_time is not None:
-                result.inter_chunk_s.append(now - last_audio_time)
-            last_audio_time = now
-            result.audio_chunk_count += 1
-
-    async for chunk in response.content.iter_any():
-        buffer.extend(chunk)
-        while b"\n" in buffer:
-            index = buffer.index(b"\n")
-            raw_line = bytes(buffer[:index])
-            del buffer[: index + 1]
-            consume_line(raw_line)
-
-    if buffer.strip():
-        consume_line(bytes(buffer))
-
-    result.text = "".join(text_parts)
-    if expect_audio and result.audio_ttfp_s is None:
-        raise ValueError("No audio chunks received from streaming response")
-    return usage
 
 
 def build_mmmu_result_records(
